requirements.txt
apify-client
firebase-admin
python-dotenv